refactor(lib-versioning): extract git remote operations from the factory
