+++ libs/versioning/src/git/factory.ts
@@ -9,8 +9,7 @@ import type { GitOperationState } from './operations/operation-state'
 import type { ListTagsOptions } from './operations/query-tags'
 import type { StageOptions, DiscardChangesOptions } from './operations/stage'
 import type { RepositoryStatus } from './operations/status'
-import { execFileSync } from 'node:child_process'
-import { createGitRef } from './models/ref'
+import type { GitCommandOptions, GitFetchOptions, GitPushOptions } from './remote'
 import { commit, amendCommit, createEmptyCommit } from './operations/commit'
 import { getChangedFilesBetween, getChangedFilesBetweenWithStatus, getCommitWithFiles } from './operations/diff'
 import { getHead, getCurrentBranch, hasUntrackedFiles } from './operations/head-info'
@@ -35,40 +34,7 @@ import {
   getModifiedFiles,
   getUntrackedFiles,
 } from './operations/status'
-
-/**
- * Validates a remote name to prevent command injection attacks.
- *
- * Git commands like fetch, pull, and push can execute arbitrary commands
- * if a malicious remote name starting with '--' is passed (e.g., '--upload-pack=cmd').
- *
- * Allowed characters: a-z, A-Z, 0-9, underscore, hyphen (not first), dot
- *
- * @param remote - The remote name to validate
- * @returns True if the remote name is safe
- * @see https://cwe.mitre.org/data/definitions/88.html (Argument Injection)
- */
-function isValidRemoteName(remote: string): boolean {
-  if (!remote || typeof remote !== 'string') {
-    return false
-  }
-
-  for (let i = 0; i < remote.length; i++) {
-    const code = remote.charCodeAt(i)
-
-    if ((code >= 97 && code <= 122) || (code >= 65 && code <= 90) || (code >= 48 && code <= 57) || code === 95) {
-      continue
-    }
-
-    if (i > 0 && (code === 45 || code === 46)) {
-      continue
-    }
-
-    return false
-  }
-
-  return true
-}
+import { getRefs, getBranches, getRemoteBranches, fetch, pull, push, getRemoteUrl } from './remote'
 
 /**
  * Git client configuration.
@@ -82,36 +48,6 @@ export interface GitClientConfig {
 
   /** Whether to throw on errors */
   readonly throwOnError?: boolean
-}
-
-/**
- * Options for git fetch operations.
- */
-export interface GitFetchOptions {
-  /** Whether to prune deleted remote branches */
-  prune?: boolean
-  /** Whether to fetch tags */
-  tags?: boolean
-}
-
-/**
- * Options for git push operations.
- */
-export interface GitPushOptions {
-  /** Whether to force push */
-  force?: boolean
-  /** Whether to set upstream tracking */
-  setUpstream?: boolean
-}
-
-/**
- * Internal options for git command execution.
- */
-interface GitCommandOptions {
-  /** Working directory for the git command */
-  cwd: string
-  /** Command timeout in milliseconds */
-  timeout: number
 }
 
 /**
@@ -495,213 +431,5 @@ export function createGitClient(config: GitClientConfig = {}): GitClient {
     pull: (remote, branch) => pull(opts, remote, branch),
     push: (remote, branch, options) => push(opts, remote, branch, options),
     getRemoteUrl: (remoteName) => getRemoteUrl(opts, remoteName),
-  }
-}
-
-/**
- * Gets all refs from the repository.
- *
- * @param options - Configuration object containing cwd and timeout
- * @param options.cwd - Working directory for the git command
- * @param options.timeout - Command timeout in milliseconds
- * @returns Array of GitRef objects representing all refs in the repository
- */
-function getRefs(options: GitCommandOptions): readonly GitRef[] {
-  try {
-    const output = execFileSync('git', ['show-ref'], {
-      encoding: 'utf-8',
-      cwd: options.cwd,
-      timeout: options.timeout,
-      stdio: ['pipe', 'pipe', 'pipe'],
-    })
-
-    const refs: GitRef[] = []
-    const lines = output.split('\n')
-
-    for (const line of lines) {
-      const trimmed = line.trim()
-      if (!trimmed) continue
-
-      const spaceIndex = trimmed.indexOf(' ')
-      if (spaceIndex === -1) continue
-
-      const hash = trimmed.slice(0, spaceIndex)
-      const fullName = trimmed.slice(spaceIndex + 1)
-
-      refs.push(createGitRef({ fullName, commitHash: hash }))
-    }
-
-    return refs
-  } catch {
-    return []
-  }
-}
-
-/**
- * Gets local branches.
- *
- * @param options - Configuration object containing cwd and timeout
- * @param options.cwd - Working directory for the git command
- * @param options.timeout - Command timeout in milliseconds
- * @returns Array of GitRef objects representing local branches
- */
-function getBranches(options: GitCommandOptions): readonly GitRef[] {
-  const refs = getRefs(options)
-  return refs.filter((ref) => ref.type === 'branch')
-}
-
-/**
- * Gets remote branches.
- *
- * @param options - Configuration object containing cwd and timeout
- * @param options.cwd - Working directory for the git command
- * @param options.timeout - Command timeout in milliseconds
- * @param remote - Optional remote name to filter branches by
- * @returns Array of GitRef objects representing remote branches
- */
-function getRemoteBranches(options: GitCommandOptions, remote?: string): readonly GitRef[] {
-  const refs = getRefs(options)
-  return refs.filter((ref) => {
-    if (ref.type !== 'remote') return false
-    if (remote && ref.remote !== remote) return false
-    return true
-  })
-}
-
-/**
- * Fetches from remote.
- *
- * @param options - Configuration object containing cwd and timeout
- * @param options.cwd - Working directory for the git command
- * @param options.timeout - Command timeout in milliseconds
- * @param remote - Remote name to fetch from (defaults to 'origin')
- * @param fetchOptions - Additional fetch configuration
- * @param fetchOptions.prune - Whether to prune deleted remote branches
- * @param fetchOptions.tags - Whether to fetch tags
- * @returns True if fetch succeeded
- */
-function fetch(options: GitCommandOptions, remote = 'origin', fetchOptions?: GitFetchOptions): boolean {
-  if (!isValidRemoteName(remote)) {
-    return false
-  }
-  const args: string[] = ['fetch', '--', remote]
-
-  if (fetchOptions?.prune) {
-    args.push('--prune')
-  }
-  if (fetchOptions?.tags) {
-    args.push('--tags')
-  }
-
-  try {
-    execFileSync('git', args, {
-      encoding: 'utf-8',
-      cwd: options.cwd,
-      timeout: options.timeout * 3,
-      stdio: ['pipe', 'pipe', 'pipe'],
-    })
-    return true
-  } catch {
-    return false
-  }
-}
-
-/**
- * Pulls from remote.
- *
- * @param options - Configuration object containing cwd and timeout
- * @param options.cwd - Working directory for the git command
- * @param options.timeout - Command timeout in milliseconds
- * @param remote - Remote name to pull from (defaults to 'origin')
- * @param branch - Optional branch name to pull
- * @returns True if pull succeeded
- */
-function pull(options: GitCommandOptions, remote = 'origin', branch?: string): boolean {
-  if (!isValidRemoteName(remote)) {
-    return false
-  }
-  const args: string[] = ['pull', '--', remote]
-  if (branch) {
-    args.push(branch)
-  }
-
-  try {
-    execFileSync('git', args, {
-      encoding: 'utf-8',
-      cwd: options.cwd,
-      timeout: options.timeout * 3,
-      stdio: ['pipe', 'pipe', 'pipe'],
-    })
-    return true
-  } catch {
-    return false
-  }
-}
-
-/**
- * Pushes to remote.
- *
- * @param options - Configuration object containing cwd and timeout
- * @param options.cwd - Working directory for the git command
- * @param options.timeout - Command timeout in milliseconds
- * @param remote - Remote name to push to (defaults to 'origin')
- * @param branch - Optional branch name to push
- * @param pushOptions - Additional push configuration
- * @param pushOptions.force - Whether to force push
- * @param pushOptions.setUpstream - Whether to set upstream tracking
- * @returns True if push succeeded
- */
-function push(options: GitCommandOptions, remote = 'origin', branch?: string, pushOptions?: GitPushOptions): boolean {
-  if (!isValidRemoteName(remote)) {
-    return false
-  }
-  const args: string[] = ['push', remote]
-
-  if (branch) {
-    args.push(branch)
-  }
-  if (pushOptions?.force) {
-    args.push('--force')
-  }
-  if (pushOptions?.setUpstream) {
-    args.push('--set-upstream')
-  }
-
-  try {
-    execFileSync('git', args, {
-      encoding: 'utf-8',
-      cwd: options.cwd,
-      timeout: options.timeout * 3,
-      stdio: ['pipe', 'pipe', 'pipe'],
-    })
-    return true
-  } catch {
-    return false
-  }
-}
-
-/**
- * Gets the URL of a remote.
- *
- * @param options - Configuration object containing cwd and timeout
- * @param options.cwd - Working directory for the git command
- * @param options.timeout - Command timeout in milliseconds
- * @param remoteName - Name of the remote (defaults to 'origin')
- * @returns The remote URL, or null if not found
- */
-function getRemoteUrl(options: GitCommandOptions, remoteName = 'origin'): string | null {
-  if (!isValidRemoteName(remoteName)) {
-    return null
-  }
-  try {
-    const output = execFileSync('git', ['remote', 'get-url', remoteName], {
-      encoding: 'utf-8',
-      cwd: options.cwd,
-      timeout: options.timeout,
-      stdio: ['pipe', 'pipe', 'pipe'],
-    })
-    return output.trim() || null
-  } catch {
-    return null
   }
 }
